[TASK] ADWD-1507 Fix column filtering for tables
There seems to be an issue with DataTables: If a row contains an empty value,
the filter is shifted that many rows to the left as there are empty values
in this row. Very strange.

Workaround is to replace all empty values with a space before rendering and
trim the values afterwards. Not beautiful, but working.